requirements.txt
streamlit
openai
python-dotenv 
pydub
numpy